+++ Intuit.TSheets.Tests/Unit/Api/DataService_CustomFieldsTests.cs
@@ -19,6 +19,7 @@
 
 namespace Intuit.TSheets.Tests.Unit.Api
 {
+    using System.Collections.Generic;
     using System.Threading.Tasks;
     using Intuit.TSheets.Client.Core;
     using Intuit.TSheets.Model;
@@ -33,6 +34,85 @@ namespace Intuit.TSheets.Tests.Unit.Api
         {
             Active = TristateChoice.Both
         };
+
+        private static readonly CustomField DummyEntity = new CustomField();
+        private static readonly List<CustomField> DummyEntities = new List<CustomField> { DummyEntity };
+
+        #region Create Method Tests
+
+        [TestMethod, TestCategory("Unit")]
+        public void CreateCustomFields_TestWithoutOptions()
+        {
+            ExpectCreate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                ApiService.CreateCustomFields(DummyEntities));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public void CreateCustomFields_TestWithOptions()
+        {
+            ExpectCreate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                ApiService.CreateCustomFields(DummyEntities));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public void CreateCustomField_TestWithoutOptions()
+        {
+            ExpectCreate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                ApiService.CreateCustomField(DummyEntity));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public void CreateCustomField_TestWithOptions()
+        {
+            ExpectCreate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                ApiService.CreateCustomField(DummyEntity));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public async Task CreateCustomFields_TestWithoutOptionsAsync()
+        {
+            ExpectCreate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                await ApiService.CreateCustomFieldsAsync(DummyEntities).ConfigureAwait(false));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public async Task CreateCustomFields_TestWithOptionsAsync()
+        {
+            ExpectCreate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                await ApiService.CreateCustomFieldsAsync(DummyEntities).ConfigureAwait(false));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public async Task CreateCustomField_TestWithoutOptionsAsync()
+        {
+            ExpectCreate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                await ApiService.CreateCustomFieldAsync(DummyEntity).ConfigureAwait(false));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public async Task CreateCustomField_TestWithOptionsAsync()
+        {
+            ExpectCreate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                await ApiService.CreateCustomFieldAsync(DummyEntity).ConfigureAwait(false));
+        }
+
+        #endregion
 
         #region Get Method Tests
 
@@ -113,6 +193,82 @@ namespace Intuit.TSheets.Tests.Unit.Api
 
             VerifyResult(
                 await ApiService.GetCustomFieldsAsync(DummyFilter, DummyRequestOptions).ConfigureAwait(false));
+        }
+
+        #endregion
+
+        #region Update Method Tests
+
+        [TestMethod, TestCategory("Unit")]
+        public void UpdateCustomFields_TestWithoutOptions()
+        {
+            ExpectUpdate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                ApiService.UpdateCustomFields(DummyEntities));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public void UpdateCustomFields_TestWithOptions()
+        {
+            ExpectUpdate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                ApiService.UpdateCustomFields(DummyEntities));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public void UpdateCustomField_TestWithoutOptions()
+        {
+            ExpectUpdate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                ApiService.UpdateCustomField(DummyEntity));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public void UpdateCustomField_TestWithOptions()
+        {
+            ExpectUpdate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                ApiService.UpdateCustomField(DummyEntity));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public async Task UpdateCustomFields_TestWithoutOptionsAsync()
+        {
+            ExpectUpdate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                await ApiService.UpdateCustomFieldsAsync(DummyEntities).ConfigureAwait(false));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public async Task UpdateCustomFields_TestWithOptionsAsync()
+        {
+            ExpectUpdate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                await ApiService.UpdateCustomFieldsAsync(DummyEntities).ConfigureAwait(false));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public async Task UpdateCustomField_TestWithoutOptionsAsync()
+        {
+            ExpectUpdate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                await ApiService.UpdateCustomFieldAsync(DummyEntity).ConfigureAwait(false));
+        }
+
+        [TestMethod, TestCategory("Unit")]
+        public async Task UpdateCustomField_TestWithOptionsAsync()
+        {
+            ExpectUpdate<CustomField>(EndpointName.CustomFields);
+
+            VerifyResult(
+                await ApiService.UpdateCustomFieldAsync(DummyEntity).ConfigureAwait(false));
         }
 
         #endregion