+++ Intuit.TSheets/Api/DataService_CustomFields.cs
@@ -19,7 +19,9 @@
 
 namespace Intuit.TSheets.Api
 {
+    using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
     using Intuit.TSheets.Client.Core;
@@ -36,6 +38,145 @@ namespace Intuit.TSheets.Api
     /// </remarks>  
     public partial class DataService
     {
+        #region Create methods
+
+        /// <summary>
+        /// Create Custom Fields.
+        /// </summary>
+        /// <remarks>
+        /// Create a single <see cref="CustomField"/> object.
+        /// </remarks>
+        /// <param name="customField">
+        /// The <see cref="CustomField"/> object to be created.
+        /// </param>
+        /// <returns>
+        /// The <see cref="CustomField"/> object that was created, along with an output
+        /// instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public (CustomField, ResultsMeta) CreateCustomField(
+            CustomField customField)
+        {
+            (IList<CustomField> customFields, ResultsMeta resultsMeta) =
+                CreateCustomFields(new[] { customField });
+
+            return (customFields.FirstOrDefault(), resultsMeta);
+        }
+
+        /// <summary>
+        /// Create Custom Fields.
+        /// </summary>
+        /// <remarks>
+        /// Create one or more <see cref="CustomField"/> objects.
+        /// </remarks>
+        /// <param name="customFields">
+        /// The set of <see cref="CustomField"/> objects to be created.
+        /// </param>
+        /// <returns>
+        /// The set of the <see cref="CustomField"/> objects that were created, along with
+        /// an output instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public (IList<CustomField>, ResultsMeta resultsMeta) CreateCustomFields(
+            IEnumerable<CustomField> customFields)
+        {
+            return AsyncUtil.RunSync(() => CreateCustomFieldsAsync(customFields));
+        }
+
+        /// <summary>
+        /// Asynchronously Create Custom Fields.
+        /// </summary>
+        /// <remarks>
+        /// Create a single <see cref="CustomField"/> object.
+        /// </remarks>
+        /// <param name="customField">
+        /// The <see cref="CustomField"/> object to be created.
+        /// </param>
+        /// <returns>
+        /// The <see cref="CustomField"/> object that was created, along with an output
+        /// instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public async Task<(CustomField, ResultsMeta)> CreateCustomFieldAsync(
+            CustomField customField)
+        {
+            (IList<CustomField> customFields, ResultsMeta resultsMeta) =
+                await CreateCustomFieldsAsync(new[] { customField }, default).ConfigureAwait(false);
+
+            return (customFields.FirstOrDefault(), resultsMeta);
+        }
+
+        /// <summary>
+        /// Asynchronously Create Custom Fields, with support for cancellation.
+        /// </summary>
+        /// <remarks>
+        /// Create a single <see cref="CustomField"/> object.
+        /// </remarks>
+        /// <param name="customField">
+        /// The <see cref="CustomField"/> object to be created.
+        /// </param>
+        /// <param name="cancellationToken">
+        /// A cancellation token that can be used by other objects or threads to receive notice of cancellation.
+        /// </param>
+        /// <returns>
+        /// The <see cref="CustomField"/> object that was created, along with an output
+        /// instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public async Task<(CustomField, ResultsMeta)> CreateCustomFieldAsync(
+            CustomField customField,
+            CancellationToken cancellationToken)
+        {
+            (IList<CustomField> customFields, ResultsMeta resultsMeta) =
+                await CreateCustomFieldsAsync(new[] { customField }, cancellationToken).ConfigureAwait(false);
+
+            return (customFields.FirstOrDefault(), resultsMeta);
+        }
+
+        /// <summary>
+        /// Asynchronously Create Custom Fields.
+        /// </summary>
+        /// <remarks>
+        /// Add one or more <see cref="CustomField"/> objects.
+        /// </remarks>
+        /// <param name="customFields">
+        /// The set of <see cref="CustomField"/> objects to be created.
+        /// </param>
+        /// <returns>
+        /// The set of the <see cref="CustomField"/> objects that were created, along with
+        /// an output instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public async Task<(IList<CustomField>, ResultsMeta resultsMeta)> CreateCustomFieldsAsync(
+            IEnumerable<CustomField> customFields)
+        {
+            return await CreateCustomFieldsAsync(customFields, default).ConfigureAwait(false);
+        }
+
+        /// <summary>
+        /// Asynchronously Create Custom Fields, with support for cancellation.
+        /// </summary>
+        /// <remarks>
+        /// Add one or more <see cref="CustomField"/> objects.
+        /// </remarks>
+        /// <param name="customFields">
+        /// The set of <see cref="CustomField"/> objects to be created.
+        /// </param>
+        /// <param name="cancellationToken">
+        /// A cancellation token that can be used by other objects or threads to receive notice of cancellation.
+        /// </param>
+        /// <returns>
+        /// The set of the <see cref="CustomField"/> objects that were created, along with
+        /// an output instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public async Task<(IList<CustomField>, ResultsMeta resultsMeta)> CreateCustomFieldsAsync(
+            IEnumerable<CustomField> customFields,
+            CancellationToken cancellationToken)
+        {
+            var context = new CreateContext<CustomField>(EndpointName.CustomFields, customFields);
+
+            await ExecuteOperationAsync(context, cancellationToken).ConfigureAwait(false);
+
+            return (context.Results.Items, context.ResultsMeta);
+        }
+
+        #endregion
+
         #region Get Methods
 
         /// <summary>
@@ -292,6 +433,145 @@ namespace Intuit.TSheets.Api
             CancellationToken cancellationToken)
         {
             var context = new GetContext<CustomField>(EndpointName.CustomFields, filter, options);
+
+            await ExecuteOperationAsync(context, cancellationToken).ConfigureAwait(false);
+
+            return (context.Results.Items, context.ResultsMeta);
+        }
+
+        #endregion
+
+        #region Update methods
+
+        /// <summary>
+        /// Update Custom Fields.
+        /// </summary>
+        /// <remarks>
+        /// Update a single <see cref="CustomField"/> object.
+        /// </remarks>
+        /// <param name="customField">
+        /// The <see cref="CustomField"/> object to be updated.
+        /// </param>
+        /// <returns>
+        /// The <see cref="CustomField"/> object that was updated, along with an output
+        /// instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public (CustomField, ResultsMeta) UpdateCustomField(
+            CustomField customField)
+        {
+            (IList<CustomField> customFields, ResultsMeta resultsMeta) =
+                UpdateCustomFields(new[] { customField });
+
+            return (customFields.FirstOrDefault(), resultsMeta);
+        }
+
+        /// <summary>
+        /// Update Custom Fields.
+        /// </summary>
+        /// <remarks>
+        /// Update one or more <see cref="CustomField"/> objects.
+        /// </remarks>
+        /// <param name="customFields">
+        /// The set of <see cref="CustomField"/> objects to be updated.
+        /// </param>
+        /// <returns>
+        /// The set of the <see cref="CustomField"/> objects that were updated, along with
+        /// an output instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public (IList<CustomField>, ResultsMeta resultsMeta) UpdateCustomFields(
+            IEnumerable<CustomField> customFields)
+        {
+            return AsyncUtil.RunSync(() => UpdateCustomFieldsAsync(customFields));
+        }
+
+        /// <summary>
+        /// Asynchronously Update Custom Fields.
+        /// </summary>
+        /// <remarks>
+        /// Add a single <see cref="CustomField"/> object.
+        /// </remarks>
+        /// <param name="customField">
+        /// The <see cref="CustomField"/> object to be updated.
+        /// </param>
+        /// <returns>
+        /// The <see cref="CustomField"/> object that was updated, along with an output
+        /// instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public async Task<(CustomField, ResultsMeta)> UpdateCustomFieldAsync(
+            CustomField customField)
+        {
+            (IList<CustomField> customFields, ResultsMeta resultsMeta) =
+                await UpdateCustomFieldsAsync(new[] { customField }, default).ConfigureAwait(false);
+
+            return (customFields.FirstOrDefault(), resultsMeta);
+        }
+
+        /// <summary>
+        /// Asynchronously Update Custom Fields, with support for cancellation.
+        /// </summary>
+        /// <remarks>
+        /// Add a single <see cref="CustomField"/> object.
+        /// </remarks>
+        /// <param name="customField">
+        /// The <see cref="CustomField"/> object to be updated.
+        /// </param>
+        /// <param name="cancellationToken">
+        /// A cancellation token that can be used by other objects or threads to receive notice of cancellation.
+        /// </param>
+        /// <returns>
+        /// The <see cref="CustomField"/> object that was updated, along with an output
+        /// instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public async Task<(CustomField, ResultsMeta)> UpdateCustomFieldAsync(
+            CustomField customField,
+            CancellationToken cancellationToken)
+        {
+            (IList<CustomField> customFields, ResultsMeta resultsMeta) =
+                await UpdateCustomFieldsAsync(new[] { customField }, cancellationToken).ConfigureAwait(false);
+
+            return (customFields.FirstOrDefault(), resultsMeta);
+        }
+
+        /// <summary>
+        /// Asynchronously Update Custom Fields.
+        /// </summary>
+        /// <remarks>
+        /// Update one or more <see cref="CustomField"/> objects.
+        /// </remarks>
+        /// <param name="customFields">
+        /// The set of <see cref="CustomField"/> objects to be updated.
+        /// </param>
+        /// <returns>
+        /// The set of the <see cref="CustomField"/> objects that were updated, along with
+        /// an output instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public async Task<(IList<CustomField>, ResultsMeta resultsMeta)> UpdateCustomFieldsAsync(
+            IEnumerable<CustomField> customFields)
+        {
+            return await UpdateCustomFieldsAsync(customFields, default).ConfigureAwait(false);
+        }
+
+        /// <summary>
+        /// Asynchronously Update Custom Fields, with support for cancellation.
+        /// </summary>
+        /// <remarks>
+        /// Update one or more <see cref="CustomField"/> objects.
+        /// </remarks>
+        /// <param name="customFields">
+        /// The set of <see cref="CustomField"/> objects to be updated.
+        /// </param>
+        /// <param name="cancellationToken">
+        /// A cancellation token that can be used by other objects or threads to receive notice of cancellation.
+        /// </param>
+        /// <returns>
+        /// The set of the <see cref="CustomField"/> objects that were updated, along with
+        /// an output instance of the <see cref="ResultsMeta"/> class containing additional data.
+        /// </returns>
+        public async Task<(IList<CustomField>, ResultsMeta resultsMeta)> UpdateCustomFieldsAsync(
+            IEnumerable<CustomField> customFields,
+            CancellationToken cancellationToken)
+        {
+            var context = new UpdateContext<CustomField>(EndpointName.CustomFields, customFields);
 
             await ExecuteOperationAsync(context, cancellationToken).ConfigureAwait(false);